refactor(reexecute/c): export CreateBlockChanFromLevelDB
